Fix noisy coeffs: Make irrToRad texture numerically stable

+++ CandelaDXR/Mathematics/Utils.h
@@ -16,11 +16,11 @@ namespace candela::mathematics
     // Integrals
 	
 	// Integral cos(theta) * cos(theta_area)/r^2 dA - for special case
-	float f1(float x, float y, float z);
+	double f1(double x, double y, double z);
 	float f1Definite(float x0, float x1, float y0, float y1, float z);
 
 	// Integral cos(theta_area)/r^2 dA - for special case
-	float f2(float x, float y, float z);
+	double f2(double x, double y, double z);
 	float f2Definite(float x0, float x1, float y0, float y1, float z);
 
 	Vector3 QuaternionToRotationXYZ(const Vector& quaternion);
